pyproject.toml
[project]
name = "uav_coordinator"
version = "0.1.0"
description = "Unified ROS package for drone coordination, position tracking, camera data, and point cloud processing"
requires-python = ">=3.12"
dependencies = [
    "ros-rclpy>=9.1.2",
    "ros-std-msgs>=5.5.1",
    "ros-geometry-msgs>=5.5.1",
    "ros-sensor-msgs>=5.5.1",
    "rclpy-intellisense>=1.2.0",
]
license = {text = "TODO: License declaration"}

[project.scripts]
position_node = "uav_coordinator.position.__main__:main"

[tool.colcon-uv-ros]
name = "uav_coordinator"

[tool.colcon-uv-ros.dependencies]
depend = ["ros-rclpy"]
exec_depend = ["ros-rclpy", "std_msgs", "geometry_msgs", "sensor_msgs"]
test_depend = ["pytest", "pytest-cov"]

[tool.colcon-uv-ros.data-files]
"share/ament_index/resource_index/packages" = ["resource/uav_coordinator"]
"share/uav_coordinator" = ["package.xml", "launch", "config"] 

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/uav_coordinator"]
